% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOrgs_add_icb_code.R
\name{cleanOrgs_add_icb_code}
\alias{cleanOrgs_add_icb_code}
\title{Adds the ICB code associated with the organisation}
\usage{
cleanOrgs_add_icb_code(df, code_column)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{code_column}{The name of the column in \code{df} that contains the lookup codes.}
}
\value{
A data frame with an additional column containing the ICB codes.
}
\description{
This function takes a dataframe and a column containing organisation codes,
then creates a new column with the ICB code based on a predefined named vector.
The function works is limited to the South East Region
}
