% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOrgs_add_shortname.R
\name{cleanOrgs_add_shortname}
\alias{cleanOrgs_add_shortname}
\title{Add short org names}
\usage{
cleanOrgs_add_shortname(df, code_column, new_col_name)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{code_column}{The name of the column in \code{df} that contains the lookup codes.}

\item{new_col_name}{The name for the new column to be created with shortened organization names.}
}
\value{
A data frame with an additional column containing the shortened organization names.
}
\description{
This function takes a dataframe and a column containing organisation codes,
then creates a new column with shortened organization names based on a predefined named vector.
The function works on both ICB and provider codes, but only for the South East region.
The function requires the user to provide a new column name so that it can be applied more than
once to the same dataframe without conflicts (allowing for separate provider and ICB columns)
}
\examples{
df <- data.frame(org_code = c('QU9', 'RHW', 'QRL'))
df <- cleanOrgs_add_shortname(df, 'org_code', 'short_org_name')
print(df)
}
