% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOrgs_create_ods_table.R
\name{cleanOrgs_create_ods_table}
\alias{cleanOrgs_create_ods_table}
\title{Create ODS table}
\usage{
cleanOrgs_create_ods_table(json_list)
}
\arguments{
\item{json_list}{an ODS json file}
}
\description{
This function takes in a list of json files extracted from
the Organisational Data Service API such as the one created by cached_ods_json_list.R
It returns a tibble. The tibble has three fields:
The organisation code,
The current organisation name as registered with the ODS
The successor code for the organisation, if that code has been retired and replaced or merged
}
