% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_cleanOrgs_get_json.R
\name{cleanOrgs_get_json}
\alias{cleanOrgs_get_json}
\title{cleanOrgs_get_json}
\usage{
cleanOrgs_get_json(df, code_column)
}
\arguments{
\item{df}{the dataframe that has the column containing organisation codes}

\item{code_column}{the column within the dataframe that has the organisation codes}
}
\description{
This is a memoised wrapper for an internal function that takes a column of NHS organisation codes
and returns a list of JSON objects extracted from the Organisational Data Service's API.
Memoise is used to cache the results of This is an internal helper function that gets used by clean_orgs_get_json.R, which
wraps this function with memoise caching the result of the API call in memory.
Caching prevents multiple repeated calls to the API,
reduces demand on the API itself and
speeds up the function when used repeatedly.
}
