% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_cleanOrgs_json_extraction.R
\name{cleanOrgs_json_extraction}
\alias{cleanOrgs_json_extraction}
\title{Extract ODS json list}
\usage{
cleanOrgs_json_extraction(df, code_column)
}
\arguments{
\item{df}{the dataframe that has the column containing organisation codes}

\item{code_column}{the column within the dataframe that has the organisation codes}
}
\description{
This function takes a column of NHS organisation codes and returns a list of
JSON objects extracted from the Organisational Data Service's API.
This is an internal helper function that gets used by clean_orgs_get_json.R, which
wraps this function with memoise to cache results.
Caching prevents multiple repeated calls to the API,
reduces demand on the API itself and
speeds up the function when used repeatedly.
}
\keyword{internal}
