% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOrgs_replace_succeeded.R
\name{cleanOrgs_replace_succeeded}
\alias{cleanOrgs_replace_succeeded}
\title{Replace with successor}
\usage{
cleanOrgs_replace_succeeded(df, code_column, verbose = TRUE)
}
\arguments{
\item{df}{The dataframe}

\item{code_column}{A string of the column that has the organisation codes being evaluated}

\item{verbose}{Logical. If TRUE, prints the number of replacements made.

This function will work on a code containing any organisation code registered in the ODS
It will not work on site codes}
}
\description{
This function takes in a dataframe with a column of organisation codes.
Uses those codes to pull out a json object for each code from the ODS API. (cached_ods_json_list)
Create a table from that API with the organisation code, organisation name and any successor code. (create_ods_table)
Compares the organisation codes in the original dataframe's code column to the table created above.
If the code has been superseded the code is replaced with the successor organisation otherwise
it remains the same.
}
