% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOrgs_update_names.R
\name{cleanOrgs_update_names}
\alias{cleanOrgs_update_names}
\title{Update organisation names}
\usage{
cleanOrgs_update_names(df, code_column, name_column, verbose = TRUE)
}
\arguments{
\item{df}{A dataframe containing organisation codes.}

\item{code_column}{A string. The name of the column in \code{df} that contains organisation codes.}

\item{name_column}{A string. The name of the column to be added/overwritten with organisation names.}

\item{verbose}{Logical. If TRUE, prints the number of unrecognised codes.}
}
\value{
A dataframe identical to \code{df}, with an additional or updated \code{name_column}.
}
\description{
This function adds or replaces a column in a dataframe with the current
organisation name as recorded by the Organisation Data Service (ODS).
It uses an organisation code column in the dataframe, retrieves JSON data
for each code using the ODS API (\code{cached_ods_json_list()}), and constructs
a lookup table via \code{create_ods_table()}.
}
\details{
Note: This function does not merge or update superseded trusts and will use
their most recent recorded name. To account for mergers, use
\code{cleanOrgs_replace_succeeded()} before calling this function.
}
